back/requirements.txt
s3transfer~=0.6.1
jmespath~=1.0.1
six~=1.16.0
python-dateutil~=2.8.2
urllib3~=1.26.15
filetype~=1.2.0
setuptools~=65.5.0
PyJWT~=2.7.0